requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langgraph==0.2.14
langgraph-checkpoint==1.0.8
langsmith==0.1.114
llama-index-agent-openai==0.3.0
llama-index-core==0.11.6
llama-index-embeddings-azure-openai==0.2.5
llama-index-embeddings-openai==0.2.4
llama-index-llms-azure-openai==0.2.0
llama-index-llms-openai==0.2.2
llama-index-vector-stores-azureaisearch==0.2.1
marshmallow==3.22.0
matplotlib-inline==0.1.7
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.7.24
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.2
tavily-python==0.4.0
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.11
zipp==3.20.1
